refactor: 예약과 예약 대기 도메인의 중복제거를 위한 Base Entity 적용
