+++ src/main/java/roomescape/domain/Reservation.java
@@ -1,49 +1,18 @@
 package roomescape.domain;
 
-import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 
 @Entity
-public class Reservation {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(nullable = false)
-    private Member member;
-
-    @Embedded
-    private Schedule schedule;
+public class Reservation extends ReservationBase {
 
     protected Reservation() {
     }
 
     public Reservation(final Member member, final Schedule schedule) {
-        this.member = member;
-        this.schedule = schedule;
+        super(member, schedule);
     }
 
     public boolean isSameTime(ReservationTime reservationTime) {
         return schedule.isSameTime(reservationTime);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public Member getMember() {
-        return member;
-    }
-
-    public Schedule getSchedule() {
-        return schedule;
     }
 }
