+++ src/main/java/roomescape/domain/ReservationBase.java
@@ -1,0 +1,45 @@
+package roomescape.domain;
+
+import jakarta.persistence.Embedded;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.MappedSuperclass;
+
+@MappedSuperclass
+public abstract class ReservationBase {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    protected Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(nullable = false)
+    protected Member member;
+
+    @Embedded
+    protected Schedule schedule;
+
+    protected ReservationBase() {
+    }
+
+    protected ReservationBase(final Member member, final Schedule schedule) {
+        this.member = member;
+        this.schedule = schedule;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public Member getMember() {
+        return member;
+    }
+
+    public Schedule getSchedule() {
+        return schedule;
+    }
+}
