+++ src/main/java/roomescape/domain/Waiting.java
@@ -1,45 +1,14 @@
 package roomescape.domain;
 
-import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 
 @Entity
-public class Waiting {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(nullable = false)
-    private Member member;
-
-    @Embedded
-    private Schedule schedule;
+public class Waiting extends ReservationBase{
 
     public Waiting(final Member member, final Schedule schedule) {
-        this.member = member;
-        this.schedule = schedule;
+        super(member, schedule);
     }
 
     protected Waiting() {
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public Member getMember() {
-        return member;
-    }
-
-    public Schedule getSchedule() {
-        return schedule;
     }
 }
